+++ packages/core/src/payment/strategies/braintree/index.ts
@@ -3,8 +3,6 @@ export * from './visacheckout';
 
 export { BraintreePaymentInitializeOptions } from './braintree-payment-options';
 export { default as BraintreePaymentProcessor } from './braintree-payment-processor';
-export { default as BraintreeVenmoPaymentStrategy } from './braintree-venmo-payment-strategy';
-export { default as BraintreeVenmoInitializeOptions } from './braintree-venmo-payment-initialize-options';
 export { default as BraintreeVisaCheckoutPaymentProcessor } from './braintree-visacheckout-payment-processor';
 export { default as BraintreeSDKCreator } from './braintree-sdk-creator';
 export { default as createBraintreePaymentProcessor } from './create-braintree-payment-processor';